changing test class file name and groups
