+++ src/test/java/com/wikia/webdriver/testcases/globalnavigationbar/Layout.java
@@ -8,13 +8,14 @@ import com.wikia.webdriver.pageobjectsfactory.pageobject.globalnav.GlobalNavigat
 import com.wikia.webdriver.pageobjectsfactory.pageobject.special.SpecialWikiActivityPageObject;
 
 import org.openqa.selenium.Dimension;
+import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
 import java.util.Arrays;
 import java.util.List;
 
-@Test(groups = {"globalnavigationbar", "Loading"})
-public class Loading extends NewTestTemplate{
+@Test(groups = {"globalnavigationbar", "Layout"})
+public class Layout extends NewTestTemplate{
 
   private static final Dimension HUBS_IN_DROPDOWN_RESOLUTION = new Dimension(768, 1024);
   private static final Dimension HUBS_OUTSIDE_DROPDOWN_RESOLUTION = new Dimension(1024, 1024);
@@ -26,7 +27,7 @@ public class Loading extends NewTestTemplate{
   private final static String deWikiName = "de.gta";
   private static final Dimension HIDE_LOGO_RESOLUTION = new Dimension(1200, 720);
 
-  @Test(groups = {"globalnavigationbar_Loading_001"})
+  @Test(groups = {"globalNavigationBarIsFixedOnScroll"})
   public void globalNavigationBarIsFixedOnScroll() {
     WikiBasePageObject base = new WikiBasePageObject(driver);
     SpecialWikiActivityPageObject wikiActivity = base.openSpecialWikiActivity();
@@ -35,8 +36,8 @@ public class Loading extends NewTestTemplate{
     wikiActivity.verifyGlobalNavigation();
   }
 
-  @Test(groups = {"globalnavigationbar_Loading_002"})
-  public void DropdownContainsExpectedLinksOn1024Resolution() {
+  @Test(groups = {"dropdownContainsExpectedLinksOnResolutionTransition"})
+  public void dropdownContainsExpectedLinksOnResolutionTransition() {
     HomePageObject homePage = new HomePageObject(driver);
     GlobalNavigationPageObject globalNav = new GlobalNavigationPageObject(driver);
     homePage.openWikiPage(this.wikiURL);
@@ -52,8 +53,8 @@ public class Loading extends NewTestTemplate{
     Assertion.assertEquals(globalNav.getDropdownLinks(), EXPECTED_LINKS_SMALL_RESOLUTION);
   }
 
-  @Test(groups = {"globalnavigationbar_Loading_003"})
-  public void gameStarLogoIsNotPresentOn768WidthResolution() {
+  @Test(groups = {"gameStarLogoIsNotPresentOn768x1024WidthResolution"})
+  public void gameStarLogoIsNotPresentOn768x1024WidthResolution() {
     HomePageObject homePage = new HomePageObject(driver);
     homePage.openWikiPage(urlBuilder.getUrlForWiki(deWikiName));
     homePage.resizeWindow(HIDE_LOGO_RESOLUTION);
@@ -62,8 +63,8 @@ public class Loading extends NewTestTemplate{
                           "GameStar Logo shouldn't be visible");
   }
 
-  @Test(groups = {"globalnavigationbar_Loading_004"})
-  public void linksArePresentOn1024WidthResolution() {
+  @Test(groups = {"linksArePresentOn1024x1024Resolution"})
+  public void linksArePresentOn1024x1024Resolution() {
     HomePageObject homePage = new HomePageObject(driver);
     GlobalNavigationPageObject globalNav = new GlobalNavigationPageObject(driver);
     homePage.openWikiPage(this.wikiURL);
@@ -73,5 +74,25 @@ public class Loading extends NewTestTemplate{
 
     driver.manage().window().setSize(HUBS_IN_DROPDOWN_RESOLUTION);
     Assertion.assertFalse(globalNav.areHubsLinksVisible());
+  }
+
+  @DataProvider
+  public Object[][] getWikisWithDisabledLocalSearch() {
+    return new Object[][]{
+        {"de.wikia"},
+        {"wikia"}
+    };
+  }
+
+  @Test(
+      groups = {"localSearchIsDisabled"},
+      dataProvider = "getWikisWithDisabledLocalSearch"
+  )
+  public void localSearchIsDisabled(String wikiName) {
+    HomePageObject homePage = new HomePageObject(driver);
+    homePage.getUrl(urlBuilder.getUrlForWiki(wikiName));
+    GlobalNavigationPageObject globalNav = homePage.getGlobalNavigation();
+
+    Assertion.assertTrue(globalNav.isLocalSearchDisabled());
   }
 }
